consolidate features/manageVault/ManageVaultCollateralAllowance.tsx into features/openMultiplyVault/common/ManageVaultCollateralAllowance.tsx
