+++ features/openMultiplyVault/common/TAllVaultStages.ts
@@ -1,0 +1,4 @@
+import { ManageMultiplyVaultStage } from '../../manageMultiplyVault/manageMultiplyVault'
+import { ManageBorrowVaultStage } from '../../manageVault/manageVault'
+
+export type TAllVaultStages = ManageMultiplyVaultStage | ManageBorrowVaultStage
