+++ features/openMultiplyVault/common/TBaseManageVaultStage.ts
@@ -1,0 +1,21 @@
+export type TBaseManageVaultStage =
+  | 'proxyWaitingForConfirmation'
+  | 'proxyWaitingForApproval'
+  | 'proxyInProgress'
+  | 'proxyFailure'
+  | 'proxySuccess'
+  | 'collateralAllowanceWaitingForConfirmation'
+  | 'collateralAllowanceWaitingForApproval'
+  | 'collateralAllowanceInProgress'
+  | 'collateralAllowanceFailure'
+  | 'collateralAllowanceSuccess'
+  | 'daiAllowanceWaitingForConfirmation'
+  | 'daiAllowanceWaitingForApproval'
+  | 'daiAllowanceInProgress'
+  | 'daiAllowanceFailure'
+  | 'daiAllowanceSuccess'
+  | 'manageWaitingForConfirmation'
+  | 'manageWaitingForApproval'
+  | 'manageInProgress'
+  | 'manageFailure'
+  | 'manageSuccess'
